Start cleanining up particle system
